requirements/base.txt
numpy
django==3.2
docker==5.0.3
django-rest-swagger
djangorestframework==3.11.1
drf-yasg
django-annoying
requests==2.25.1
injector==0.18.4
bs4==0.0.1
pandas==1.2.0
psycopg2-binary==2.8.6
django-widget-tweaks==1.4.8
tqdm==4.56.0
django-extensions==3.1.0
python-binance==1.0.15
django-filter==2.4.0
more-itertools==8.8.0
django-bootstrap4==21.1
sklearn
cryptowatch==0.0.11
coinapi_rest_v1
